Refactor: 共通する機能を一つのコントローラに統合
